handson/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
line-bot-sdk==1.15.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
requests==2.23.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
